Add contexts on the server side
This makes server calls to RPC methods context-aware.

In the case of regular RPC-via-libp2p calls, a context is
created and associated to the incoming stream. If the stream
is closed or reset, the context will be cancelled. It is
up to the method's implemented to decide what to do with that.

In the case of RPC-shortcut for local calls, the context optionally
provided by the client is used directly.