Resolve possible race
When a call dies badly, we cancel the context and that attempts
to set the error. We probably should not set that error at all
since it will likely overwrite the original and at worse
cause a race. This fixes that and also adds getError() to
avoid a race altogether.