Permissioned RPC endpoints
This commit introduces authorization for rpc server.

It introduces an authorization function, which can be passed while
creating a new server. One such function is provided which uses a map
(maps "service.method" of a peer to boolean permission) to check
permissions. User can choose to use that function or a define a
new function as long as it follows the prescribed format.

It follows a whitelist approach(i.e., by default does not allow)
- If an authorization is not passed, server would not let any peer
access any methods