Combine outputs in MultiStream method
Upon some thinking, I have decided that the MultiStream method is most useful
when all the responses are combined in a single channel.

Thus I have edited the method and the signature. It did not seem worth to add
yet an extra method to keep the old one, and then maintain 2x of tricky code.

It is much easier to call a method in multiple places and just have all the
responses tickle on a single channel as they happen, than having to read on
multiple channels, which forces the clients to either reflect, or launch
separate goroutines etc.