Fix: MultiStream panics when using Receive-only channel for out
This fixes a panic when the channel passed in for receiving replies in
MultiStream() is receive-only. The underlying problem is that the channels
created as Reply channels for the multiplexed calls to Stream were of the same
type as the channel passed-in, but these channels are read from in order to
place all the items in the original "out" channel.

We had not seen this before as channels passed-in were normally bi-directional.